requirements.txt
python-telegram-bot[job-queue,webhooks]==22.7
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.0
pytz==2024.1
oauth2client==4.1.3
reportlab==4.2.5
groq==0.26.0